Missing default case
I noticed that `getSymbol` could fall through without returning
any value.  The expected return type is a string, but without this
default case, it can return `undefined`.

Also, added an extra guard to an if statement to ensure that
`contextRegex` is really defined.  I see that they are (currently)
defined together.  But the TypeScript compiler flagged this
as a potential issue.  Seems reasonable to put this guard in
place just in case at some future the "defined together" behavior
changes.