Fix build of ES5 version
This is a minefield.  Previous attempts pull in runtime dependencies
which is normally fine for node.js, but not browsers.
This commit removes babel, uses regenerator to rewrite
the generator code, which also rewrites const/let, arrow functions.
Polyfills are then added to a new file polyfill.js which gets
concatenated to the browserified output.
This now works in Nashorn (limited testing).  More testing may
reveal more functions to be polyfilled (future issue).

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>