$eval() - array inputs should be wrapped
The logic for handling the second parameter to this function should be the same as for handling the input to any expression.
This includes wrapping a top level array input in a singleton sequence so it gets treated as a single input.
This code was missing from this function and is added here.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>