Protect `__evaluate_entry` and `__evaluate_exit` callbacks. (#700)
These two internal APIs are very useful in debugging scenarios, and for imposing time/depth constraints on queries.
However, this internal API should only be accessible when configuring an expression programmatically, and not
from inside of a query. Otherwise, a query can be manipulated to remove such diagnostics or constraints.

By changing both binding keys to Symbol, they can no longer be accessed inside of the query since the Symbol API
is not accessible there.